Fix some missing typehints

+++ src/Backup/Cleaner/Stepwise.php
@@ -96,19 +96,19 @@ class Stepwise extends Abstraction implements Simulator
         $all   = new Range($start, $end, new Stepwise\Keeper\All());
 
         // define the range that keeps backups per day
-        $end   = mktime(0, 0, 0, date('m', $end), date('d', $end) - $this->daysToKeepDaily, date('Y', $end));
+        $end   = mktime(0, 0, 0, date('m', $end), (int) date('d', $end) - $this->daysToKeepDaily, date('Y', $end));
         $daily = new Range($all->getEnd(), $end, new Stepwise\Keeper\OnePerGroup('Ymd'));
 
         // define the range that keeps backups per week
-        $end    = mktime(0, 0, 0, date('m', $end), date('d', $end) - (7 * $this->weeksToKeepWeekly), date('Y', $end));
+        $end    = mktime(0, 0, 0, date('m', $end), (int) date('d', $end) - (7 * $this->weeksToKeepWeekly), date('Y', $end));
         $weekly = new Range($daily->getEnd(), $end, new Stepwise\Keeper\OnePerGroup('YW'));
 
         // define the range that keeps backups per month
-        $end     = mktime(0, 0, 0, date('m', $end) - $this->monthToKeepMonthly, date('d', $end), date('Y', $end));
+        $end     = mktime(0, 0, 0, (int) date('m', $end) - $this->monthToKeepMonthly, date('d', $end), date('Y', $end));
         $monthly = new Range($weekly->getEnd(), $end, new Stepwise\Keeper\OnePerGroup('Ym'));
 
         // define the range that keeps backups per year
-        $end    = mktime(0, 0, 0, date('m', $end), date('d', $end), date('Y', $end) - $this->yearsToKeepYearly);
+        $end    = mktime(0, 0, 0, date('m', $end), date('d', $end), (int) date('Y', $end) - $this->yearsToKeepYearly);
         $yearly = new Range($monthly->getEnd(), $end, new Stepwise\Keeper\OnePerGroup('Y'));
 
         // delete all backups older then configured year range
